App_Backend/backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.5
django-allauth==0.52.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
oauthlib==3.2.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
